split accelerated_train and helper functions
split both multi and single phase training into an accelerated_train module

this removes all instances of click from the core training functionality. This will allow us to eventually implement handlers shelling out to these various modes of training and even multi processing functionality since you can only pickle primitives in python using multiprocessing
